fix(tinygo): add sql lite build profile

+++ sql/types/geometry.go
@@ -569,7 +569,7 @@ func (t GeometryType) MatchSRID(v interface{}) error {
 }
 
 func ValidateSRID(srid int, funcName string) error {
-	if srid < 0 || srid > math.MaxUint32 {
+	if srid < 0 || uint64(srid) > math.MaxUint32 {
 		return sql.ErrInvalidSRID.New(funcName)
 	}
 	if _, ok := SupportedSRIDs[uint32(srid)]; !ok {
